SFX and Music Volume Functionality
Uses Audio Mixer to change SFX and Music volume via slider in settings menu.
