+++ Roguelite/Assets/ButtonManager.cs
@@ -1,0 +1,18 @@
+﻿using UnityEngine;
+using UnityEngine.UI;
+using Utils;
+
+public sealed class ButtonManager : MonoBehaviour
+{
+    [SerializeField] private Button[] m_Buttons;
+
+    public void DisableButtons()
+    {
+        UserInterfaceUtils.DisableButtons(m_Buttons);
+    }
+
+    public void EnableButtons()
+    {
+        UserInterfaceUtils.EnableButtons(m_Buttons);
+    }
+}
