+++ Roguelite/Assets/ButtonManager.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: d712097b90eb4771b38c7d26d717a308
+timeCreated: 1590059421
