+++ Roguelite/Assets/Scripts/UI/Equipment.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fcd11b6ac8697974dba08e86575401a4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
