+++ Roguelite/Assets/Scripts/UI/Equipment/EquipmentUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 723aa31ecd7701545a70c535fcaf18f3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
