--- Roguelite/Assets/Scripts/UI/EquipmentUI.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: fcd11b6ac8697974dba08e86575401a4
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
