--- Roguelite/Assets/Scripts/UI/EquipmentUI/EquipmentSlotUI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2904f96ff74466e4390ad3ae46fba0a7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
