--- Roguelite/Assets/Scripts/UI/EquipmentUI/EquipmentUI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 723aa31ecd7701545a70c535fcaf18f3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
