+++ Roguelite/Assets/Scripts/UI/Settings.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 38aa83c0b306d8f499f4609925b7a045
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
