+++ Roguelite/Assets/Scripts/UI/Settings/EffectsVolume.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: e38b0df7d18e4ee6b1de496bf49b34e7
+timeCreated: 1590060660
