+++ Roguelite/Assets/Scripts/UI/Settings/MusicVolume.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 00290a8a360a39e46b59426e53890755
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
